% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checkZoteroConn}
\alias{checkZoteroConn}
\title{Check the connection with Zotero library}
\usage{
checkZoteroConn(verbose = FALSE)
}
\arguments{
\item{verbose}{if TRUE print information about the Zotero connection
(default: FALSE).}
}
\value{
A logical:
\itemize{
\item TRUE if the connection can be established
\item Or FALSE if the connection cannot be established.
}
}
\description{
Check the connection with Zotero library
}
\seealso{
\link{connectToZotero}
}
