% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{forgetZoteroConn}
\alias{forgetZoteroConn}
\title{Forget a Zotero connection}
\usage{
forgetZoteroConn(connection)
}
\arguments{
\item{connection}{the id of the connection to forget.}
}
\description{
Forget a Zotero connection
}
\seealso{
\link{lsZoteroConnections}, \link{checkZoteroConn}, \link{connectToZotero}
}
