% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoteroBib.R
\name{getZoteroBib}
\alias{getZoteroBib}
\title{Get Zotero bibiliography from cloud}
\usage{
getZoteroBib(
  baseURL = "https://api.zotero.org",
  userID,
  key,
  latestVersion = NA,
  excludedTypes = "attachment",
  by = 100,
  verbose = FALSE
)
}
\arguments{
\item{userID}{Zotero user identifier}

\item{key}{Zotero key}

\item{latestVersion}{A numeric indicating the version after which
modified records should be taken (Default: NA ==> take all records)}

\item{excludedTypes}{the types of record to exclude (default: "attachment")}

\item{by}{Number of records to be taken at once (default: 100 (maximum
value for Zotero API))}

\item{verbose}{If TRUE messages regarding the download of records are
displayed.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}  with a \code{latestVersion} single numeric attribute
and the following fields:
\itemize{
\item \strong{key}: the Zotero internal key of the record
\item \strong{bib}: bibtex representation of the record
\item \strong{type}: the type of the record
\item \strong{id}: the record identifier
\item \strong{title}: the record title
\item \strong{journal}: the record journal
\item \strong{year}: publication year
\item \strong{authors}: the record authors
\item \strong{pmid}: PubMed identifier
\item \strong{doi}: Digital Object Identifier
\item \strong{url}: the record URL
}
}
\description{
Internal function (not exported)
}
