% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{lsZoteroConnections}
\alias{lsZoteroConnections}
\title{List all registered Zotero connection}
\usage{
lsZoteroConnections()
}
\description{
List all registered Zotero connection
}
\seealso{
\link{forgetZoteroConn}, \link{checkZoteroConn}, \link{connectToZotero}
}
