% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoteroBib.R
\name{mergeZoteroBib}
\alias{mergeZoteroBib}
\title{Merge 2 zoteroBib object}
\usage{
mergeZoteroBib(x, y)
}
\arguments{
\item{x}{a zoteroBib}

\item{y}{a zoteroBib}
}
\description{
Internal function (not exported)
}
