% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoteroBib.R
\name{updateZoteroBib}
\alias{updateZoteroBib}
\title{Update Zotero bibliography}
\usage{
updateZoteroBib(incremental = TRUE, verbose = TRUE)
}
\arguments{
\item{incremental}{a logical indicating if the update should be incremental
(Default: TRUE). If FALSE all the records are downloaded}

\item{verbose}{a logical indicating if additional messages should be
displayed.}
}
\description{
Update the local copy of Zotero bibliography
}
